% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{buildPlotTotalDF}
\alias{buildPlotTotalDF}
\title{Build all labels in the graph}
\usage{
buildPlotTotalDF(path, tree, ig, binVector = 1:12)
}
\arguments{
  \item{path}{path as returned from getPath() or a vector
  of two variety names which exist in ig}

  \item{tree}{the data tree (in data frame format)}

  \item{ig}{the graph representation of the data tree (in
  igraph format)}

  \item{binVector}{vector of numbers between 1 and
  length(binVector), each repeated exactly once}
}
\description{
This function takes the spreadTotalDF object (from the
buildSpreadTotalDF function) and the path object as inputs.
From these objects, it creates a data frame object of the
text label positions for the varieties in the path, as well
as the edges only in the varieties in the path.
}
\seealso{
\url{http://www.r-project.org} for iGraph information

\url{http://www.r-project.org} for iGraph information

\code{\link{getPath}} for information on input path
building
}

