% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{buildSpreadTotalDF}
\alias{buildSpreadTotalDF}
\title{Build a data frame where the varieties are spread so they do not overlap}
\usage{
buildSpreadTotalDF(geneal, ig, bin = 12)
}
\arguments{
\item{geneal}{the full genealogy  (in data frame format)}

\item{ig}{the graph representation of the data genealogy (in igraph format)}

\item{bin}{the number of bins to determine the vertical positions of nodes (default is 12). For more information on choosing bin size, please visit the ggenealogy vignette}
}
\description{
Constructs a data frame object so that varieties are spread such that they do not overlap, even
though the x-axis position will represent dates.
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

