% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getChild}
\alias{getChild}
\title{Returns the children of a particular variety (if they exist)}
\usage{
getChild(v1, geneal)
}
\arguments{
\item{v1}{the label of the vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
This function returns zero or more values that indicate the children of the inputted variety.
}
\examples{
data(sbGeneal)
getChild("Tokyo", sbGeneal)
getChild("Essex", sbGeneal)
}
\references{
Rutter L, VanderPlas S, Cook D, Graham MA (2019).
ggenealogy: An R Package for Visualizing Genealogical Data.
\emph{Journal of Statistical Software},
\bold{89}(13), 1--31.
\doi{10.18637/jss.v089.i13}
}
