% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getDate}
\alias{getDate}
\title{Determine the date of a variety}
\usage{
getDate(v1, geneal)
}
\arguments{
\item{v1}{the label of the variety/vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
Returns the documented date of the inputted variety
}
\examples{
data(sbGeneal)
getDate("Essex", sbGeneal)
getDate("Tokyo", sbGeneal)
}

