% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getEdges}
\alias{getEdges}
\title{Returns edges (vertex names and edge weights) for a tree}
\usage{
getEdges(ig, tree)
}
\arguments{
  \item{ig}{the graph representation of the data tree (in
  igraph format)}

  \item{tree}{the data tree (in data frame format)}
}
\description{
Returns a matrix, where each row contains information about
an edge (two vertex names and edge weight, if present) of
the tree.
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
getEdges(ig, sbTree)
}

