% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getNodes}
\alias{getNodes}
\title{Returns the nodes for a full genealogy}
\usage{
getNodes(geneal)
}
\arguments{
\item{geneal}{the full genealogy (in data frame format)}
}
\description{
Returns a character list, where rows contains names of the unique nodes in the full genealogy
}
\examples{
data(sbGeneal)
getNodes(sbGeneal)
}
\references{
Rutter L, VanderPlas S, Cook D, Graham MA (2019).
ggenealogy: An R Package for Visualizing Genealogical Data.
\emph{Journal of Statistical Software},
\bold{89}(13), 1--31.
\doi{10.18637/jss.v089.i13}
}
