% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getParent}
\alias{getParent}
\title{Returns the parents of a particular variety (if they exist)}
\usage{
getParent(v1, geneal)
}
\arguments{
\item{v1}{the label of the vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
This function returns up to two values that indicate the parents of the inputted variety.
}
\examples{
data(sbGeneal)
getParent("Tokyo", sbGeneal)
getParent("Essex", sbGeneal)
}
\references{
Rutter L, VanderPlas S, Cook D, Graham MA (2019).
ggenealogy: An R Package for Visualizing Genealogical Data.
\emph{Journal of Statistical Software},
\bold{89}(13), 1--31.
\doi{10.18637/jss.v089.i13}
}
