% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getVariable}
\alias{getVariable}
\title{Determine the date of a variety}
\usage{
getVariable(v1, geneal, colName)
}
\arguments{
\item{v1}{the label of the vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}

\item{colName}{the name of the column of the data frame that contains the quantitative variable of interest (in character string format)}
}
\description{
Returns the documented date of the inputted variety
}
\examples{
data(sbGeneal)
getVariable("Essex", sbGeneal, "devYear")
getVariable("Tokyo", sbGeneal, "yield")
}
\references{
Rutter L, VanderPlas S, Cook D, Graham MA (2019).
ggenealogy: An R Package for Visualizing Genealogical Data.
\emph{Journal of Statistical Software},
\bold{89}(13), 1--31.
\doi{10.18637/jss.v089.i13}
}
