% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getYear}
\alias{getYear}
\title{Determine the year of a variety}
\usage{
getYear(v1, tree)
}
\arguments{
  \item{v1}{the label of the variety/vertex of interest (in
  character string format)}

  \item{tree}{the data tree (in data frame format)}
}
\description{
Returns the documented year of the inputted variety
}
\examples{
data(sbTree)
getYear("Essex",sbTree)
getYear("Tokyo",sbTree)
}

