% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{isParent}
\alias{isParent}
\title{Determine if a variety is a parent of another}
\usage{
isParent(child, parent, tree)
}
\arguments{
  \item{child}{possible child variety}

  \item{parent}{possible parent variety}

  \item{tree}{the data tree (in data frame format)}
}
\description{
Returns a boolean variable for whether the second variety
is a parent of the first variety
}
\examples{
data(sbTree)
isParent("Essex","Young",sbTree)
isParent("Young","Essex",sbTree)
}

