% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mywidget.R
\name{mywidgetOutput}
\alias{mywidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
mywidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

