% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{nodeToDF}
\alias{nodeToDF}
\title{Returns the data frame representation of all ancestors and descendants of a variety}
\usage{
nodeToDF(tlist, branch = 0, par.id = NA, id = 1)
}
\arguments{
  \item{tlist}{list of varieties}

  \item{branch}{of particular variety in tree}

  \item{par.id}{the id of the parent}

  \item{id}{id offset}
}
\description{
Converts the list-style-tree to a data frame, where each
variety has an id value and references its' parent's id
value. ID value ranges correspond to generation. It is
possible that with more complex trees the range of id
values may need to expand to reduce the probability of two
varieties being assigned the same id value.
}

