% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mywidget.R
\name{renderMywidget}
\alias{renderMywidget}
\title{Widget render function for use in Shiny}
\usage{
renderMywidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

