% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{treeToIG}
\alias{treeToIG}
\title{Process the tree graph}
\usage{
treeToIG(tree, vertexinfo = NULL, edgeweights = 1, isDirected = FALSE)
}
\arguments{
  \item{tree}{the data tree (in data frame format)}

  \item{vertexinfo}{(default NULL) either names of columns
  in the tree which should be added to the database as
  vertex information or a data frame with information for
  all vertices such that the first column contains vertex
  names.}

  \item{edgeweights}{(default 1) name of a column which
  contains edge weights}

  \item{isDirected}{(default FALSE) should the graph be a
  directed graph?}
}
\description{
Processes the tree into an igraph object with appropriate
vertex information, graph type, and edge weights.
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

