% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{buildAncDesCoordDF}
\alias{buildAncDesCoordDF}
\title{Returns the coordinate positions of all ancestors and descendants of a variety.}
\usage{
buildAncDesCoordDF(df)
}
\arguments{
\item{df}{the data frame of the ancestors and descendants of a variety (from function buildAncDesTotalDF)}
}
\description{
Calculates coordinates to plot each ancestors and descendant of a variety in a lineage. The x and y values describe the coordinates of the label, while the xstart, ystart, xend, and yend values describe the edges of the label.
}
\seealso{
\code{\link{buildAncList}} for information on determining ancestors

\code{\link{buildDesList}} for information on determining descendants
}

