% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{buildDesList}
\alias{buildDesList}
\title{Returns the descendants of a particular variety (if they exist).}
\usage{
buildDesList(v1, geneal, gen = 0)
}
\arguments{
\item{v1}{the label of the variety/vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}

\item{gen}{the generation (note: This should be left as default, as any other input will not affect results)}
}
\description{
This function returns a nested list of the descendants of the inputted variety.
}
\examples{
data(sbGeneal)
getParent("Essex", sbGeneal)
buildDesList("Essex", sbGeneal, 3)
}
\seealso{
\code{\link{getChild}} for information on determining children
}

