% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{buildPlotTotalDF}
\alias{buildPlotTotalDF}
\title{Build all labels in the graph}
\usage{
buildPlotTotalDF(path, geneal, ig, colName, colNameY = "", bin = 12)
}
\arguments{
\item{path}{path as returned from getPath() or a vector of two variety names which exist in ig}

\item{geneal}{the full genealogy  (in data frame format)}

\item{ig}{the graph representation of the data genealogy (in igraph format)}

\item{colName}{the name of the column of the data frame that contains the quantitative variable of interest (in character string format)}

\item{colNameY}{the name of the second optional column of the data frame that contains the second optional quantitative variable of interest (in character string format). This optional quantitative variable will be plotted on the vertical axis.}

\item{bin}{the number of bins to determine the vertical positions of nodes (default is 12). For more information on choosing bin size, please visit the ggenealogy vignette}
}
\description{
This function takes the spreadTotalDF object (from the buildSpreadTotalDF function) and the path object
as inputs. From these objects, it creates a data frame object of the text label positions for the
varieties in the path, as well as the edges only in the varieties in the path.
}
\seealso{
\url{http://www.r-project.org} for iGraph information

\url{http://www.r-project.org} for iGraph information

\code{\link{getPath}} for information on input path building
}

