% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getAncestors}
\alias{getAncestors}
\title{Returns a list of the ancestors of a particular variety (if they exist)}
\usage{
getAncestors(v1, tree, gen = 3)
}
\arguments{
  \item{v1}{the label of the variety/vertex of interest (in
  character string format)}

  \item{tree}{the data tree (in data frame format)}

  \item{gen}{the number of generations back to include as
  ancestors}
}
\description{
This function returns a list of the ancestors of the
inputted variety within and including a given number of
generations
}
\examples{
data(sbTree)
getParent("Essex", sbTree)
getAncestors("Essex", sbTree, 1)
getAncestors("Essex", sbTree, 5)
}

