% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getChild}
\alias{getChild}
\title{Returns the children of a particular variety (if they exist)}
\usage{
getChild(v1, tree)
}
\arguments{
  \item{v1}{the label of the variety/vertex of interest (in
  character string format)}

  \item{tree}{the data tree (in data frame format)}
}
\description{
This function returns zero or more values that indicate the
children of the inputted variety.
}
\examples{
data(sbTree)
getChild("Tokyo", sbTree)
getChild("Essex", sbTree)
}

