% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getChild}
\alias{getChild}
\title{Returns the children of a particular variety (if they exist)}
\usage{
getChild(v1, geneal)
}
\arguments{
\item{v1}{the label of the vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
This function returns zero or more values that indicate the children of the inputted variety.
}
\examples{
data(sbGeneal)
getChild("Tokyo", sbGeneal)
getChild("Essex", sbGeneal)
}

