% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getDescendants}
\alias{getDescendants}
\title{Returns a list of the descendants of a particular variety (if they exist)}
\usage{
getDescendants(v1, geneal, gen = 3)
}
\arguments{
\item{v1}{the label of the variety/vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}

\item{gen}{the number of generations back to include as descendants}
}
\description{
This function returns a list of the descendants of the inputted variety within and including a given number of generations
}
\examples{
data(sbGeneal)
getChild("Essex", sbGeneal)
getDescendants("Essex", sbGeneal, 1)
getDescendants("Essex", sbGeneal, 3)
}

