% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getEdges}
\alias{getEdges}
\title{Returns edges (vertex names and edge weights) for the full genealogy}
\usage{
getEdges(ig, geneal)
}
\arguments{
\item{ig}{the graph representation of the data genealogy (in igraph format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
Returns a matrix, where each row contains information about an edge (two vertex names and edge weight, if present) of the full genealogy.
}
\examples{
data(sbGeneal)
ig <- dfToIG(sbGeneal)
getEdges(ig, sbGeneal)
}

