% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getNodes}
\alias{getNodes}
\title{Returns the nodes for a tree}
\usage{
getNodes(tree)
}
\arguments{
  \item{tree}{the data tree (in data frame format)}
}
\description{
Returns a character list, where roww contains names of the
unique nodes in a tree.
}
\examples{
data(sbTree)
getNodes(sbTree)
}

