% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getNodes}
\alias{getNodes}
\title{Returns the nodes for a full genealogy}
\usage{
getNodes(geneal)
}
\arguments{
\item{geneal}{the full genealogy (in data frame format)}
}
\description{
Returns a character list, where rows contains names of the unique nodes in the full genealogy
}
\examples{
data(sbGeneal)
getNodes(sbGeneal)
}

