% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getParent}
\alias{getParent}
\title{Returns the parents of a particular variety (if they exist)}
\usage{
getParent(v1, tree)
}
\arguments{
  \item{v1}{the label of the variety/vertex of interest (in
  character string format)}

  \item{tree}{the data tree (in data frame format)}
}
\description{
This function returns up to two values that indicate the
parents of the inputted variety.
}
\examples{
data(sbTree)
getParent("Tokyo", sbTree)
getParent("Essex", sbTree)
}

