% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getYear}
\alias{getYear}
\title{Determine the year of a variety}
\usage{
getYear(v1, geneal)
}
\arguments{
\item{v1}{the label of the variety/vertex of interest (in character string format)}

\item{geneal}{the full genealogy  (in data frame format)}
}
\description{
Returns the documented year of the inputted variety
}
\examples{
data(sbGeneal)
getYear("Essex", sbGeneal)
getYear("Tokyo", sbGeneal)
}

