% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{plotDegMatrix}
\alias{plotDegMatrix}
\title{Returns the image object to show the heat map of degrees between the inputted set of vertices}
\usage{
plotDegMatrix(varieties, ig, tree, xLab = "Variety", yLab = "Variety",
  legendLab = "Degree")
}
\arguments{
  \item{varieties}{subset of varieties used to generate the
  heat map}

  \item{ig}{the graph representation of the data tree (in
  igraph format)}

  \item{tree}{the data tree (in data frame format)}

  \item{xLab}{string label on the x axis (default is
  "Variety")}

  \item{yLab}{string label on the y axis (default is
  "Variety")}

  \item{legendLab}{string label on the legend (default is
  "Degree")}
}
\description{
Returns the image object to show the heat map of degrees
between the inputted set of vertices
}
\examples{
data(sbTree)
ig = treeToIG(sbTree)
varieties=c("Bedford", "Calland", "Narow", "Pella", "Tokyo", "Young", "Zane")
p = plotDegMatrix(varieties, ig, sbTree, "Soybean label", "Soybean label", "Degree")
p + ggplot2::scale_fill_continuous(low="white", high="darkgreen")
}
\seealso{
\url{http://www.r-project.org} for iGraph information
}

